+++ bootstrapper/src/main/java/ch/hsr/adv/lib/bootstrapper/ADV.java
@@ -1,9 +1,9 @@
 package ch.hsr.adv.lib.bootstrapper;
 
+import ch.hsr.adv.commons.core.logic.util.ADVException;
 import ch.hsr.adv.lib.core.logic.ADVCore;
 import ch.hsr.adv.lib.core.logic.ADVModule;
 import ch.hsr.adv.lib.core.logic.GuiceCoreModule;
-import ch.hsr.adv.lib.core.logic.util.ADVException;
 import com.google.inject.Guice;
 import com.google.inject.Injector;
 import com.google.inject.Singleton;
