--- lib-core/src/main/java/ch/hsr/adv/lib/core/logic/domain/ADVElement.java
@@ -1,53 +1,0 @@
-package ch.hsr.adv.lib.core.logic.domain;
-
-import ch.hsr.adv.commons.core.logic.domain.styles.ADVStyle;
-
-/**
- * Represents an element of a snapshot. An
- * element could for example be a node in a graph.
- *
- * @param <T> Type of content. Could either be a simple Java class like
- *            String or a more complexe class designed by the module developer.
- */
-public interface ADVElement<T> {
-    /**
-     * Gets a unique id of the element. The id only needs to be unique within
-     * one snapshot. Preferably use the same id for the same element in
-     * different snapshots.
-     *
-     * @return a snapshot-wide unique id
-     */
-    long getId();
-
-    /**
-     * Returns the style of the element. Use ADVDefaultStyle for a default
-     * implementation.
-     *
-     * @return the style of the element
-     */
-    ADVStyle getStyle();
-
-    /**
-     * Offers the user the possibility to set the position of an element.
-     * The position will be ignored by the ui if it equals 0;
-     *
-     * @return the x position of the element
-     */
-    int getFixedPosX();
-
-    /**
-     * Offers the user the possibility to set the position of an element.
-     * The position will be ignored by the ui if it equals 0;
-     *
-     * @return the y position of the element
-     */
-    int getFixedPosY();
-
-    /**
-     * The content could either be a simple Java class like String or a more
-     * complexe class designed by the module developer.
-     *
-     * @return the content of the element
-     */
-    T getContent();
-}
