+++ lib-core/src/main/java/ch/hsr/adv/lib/core/logic/util/ADVConnectionException.java
@@ -1,5 +1,7 @@
 package ch.hsr.adv.lib.core.logic.util;
 
+import ch.hsr.adv.commons.core.logic.util.ADVException;
+
 /**
  * To be thrown for any exception to do with the connection to the ADVCore UI.
  */
