+++ module-bb-array/src/test/java/ch/hsr/adv/lib/array/logic/ArrayBuilderTest.java
@@ -1,7 +1,7 @@
 package ch.hsr.adv.lib.array.logic;
 
+import ch.hsr.adv.commons.core.logic.domain.ADVElement;
 import ch.hsr.adv.lib.array.logic.domain.ArrayTestModule;
-import ch.hsr.adv.lib.core.logic.domain.ADVElement;
 import ch.hsr.adv.lib.core.logic.domain.ModuleGroup;
 import com.google.inject.Inject;
 import org.jukito.JukitoRunner;
