+++ module-bb-graph/src/main/java/ch/hsr/adv/lib/graph/logic/GraphElement.java
@@ -1,7 +1,7 @@
 package ch.hsr.adv.lib.graph.logic;
 
+import ch.hsr.adv.commons.core.logic.domain.ADVElement;
 import ch.hsr.adv.commons.core.logic.domain.styles.ADVStyle;
-import ch.hsr.adv.lib.core.logic.domain.ADVElement;
 import ch.hsr.adv.lib.graph.logic.domain.ADVVertex;
 
 /**
