+++ module-bb-graph/src/main/java/ch/hsr/adv/lib/graph/logic/domain/ADVEdge.java
@@ -1,6 +1,6 @@
 package ch.hsr.adv.lib.graph.logic.domain;
 
-import ch.hsr.adv.lib.core.logic.domain.ADVRelation;
+import ch.hsr.adv.commons.core.logic.domain.ADVRelation;
 
 /**
  * An edge always has an end and a start vertex. It can be directed or
