+++ module-bb-graph/src/test/java/ch/hsr/adv/lib/graph/logic/GraphBuilderTest.java
@@ -1,6 +1,6 @@
 package ch.hsr.adv.lib.graph.logic;
 
-import ch.hsr.adv.lib.core.logic.domain.ADVElement;
+import ch.hsr.adv.commons.core.logic.domain.ADVElement;
 import ch.hsr.adv.lib.core.logic.domain.ModuleGroup;
 import ch.hsr.adv.lib.graph.logic.domain.GraphTestModule;
 import com.google.inject.Inject;
